+++ src/header-helpers.ts
@@ -51,3 +51,48 @@ export function is(message: RequestInterface | ResponseInterface, type: string):
   return false;
 
 }
+
+type Preferences = {
+  [preference: string]: true | string;
+};
+
+/**
+ * Parses a RFC7240 Prefer header.
+ *
+ * It's a naive parser as it assumes a fairly simple subset of
+ * Prefer.
+ *
+ * TODO: Make this parse every possible variation.
+ */
+export function parsePrefer(header: string | null): Preferences {
+
+  if (!header) {
+    return {};
+  }
+
+  const result: Preferences = {};
+
+  for (const headerItem of splitHeader(header)) {
+
+    const [keyValue] = headerItem.split(';');
+    const [key, value] = keyValue.split('=');
+    result[key.toLowerCase()] = value !== undefined ? value : true;
+
+  }
+
+  return result;
+
+}
+
+/**
+ * This function takes a multi-value comma-separated header and splits it
+ * into multiple headers.
+ *
+ * TODO: In the future this function will respect comma's appearing within
+ * quotes and ignore them. It doesn't right now.
+ */
+export function splitHeader(header: string): string[] {
+
+  return header.split(',').map(item => item.trim());
+
+}